Massive format update. I'm sorry.
From now on:
1. All dates will be iso-8601 in all their leading zero glory.  If you
built a custom date parser for this, I'm so sorry.
2. I will now be transcribing and publishing the 7 day averages as
written by the MWRA.  They will be in new columns to the right of the
single day datums.
3. I will be publishing a new file, 7day-diff.tsv, which will show the
difference between my calculated 7 day averages and the published 7 day
averages.  I expect these numbers to be less than 1, though in rare
cases, they are greater than 1, for reasons that I don't really
understand.  Honestly, I don't need to publish them, but this makes it
easier on me if I do, because my workflow heavily relies on git diff.

My plotit script has been changed to produce and use of the iso-8601
format dates.